+++ src/test/kotlin/uk/gov/justice/digital/hmpps/learnerrecordsapi/integration/LearnersResourceIntTest.kt
@@ -1,7 +1,6 @@
 package uk.gov.justice.digital.hmpps.learnerrecordsapi.integration
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import net.minidev.json.JSONObject
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
@@ -11,16 +10,13 @@ import org.springframework.http.MediaType
 import software.amazon.awssdk.services.sqs.model.PurgeQueueRequest
 import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.config.HmppsBoldLrsExceptionHandler
-import uk.gov.justice.digital.hmpps.learnerrecordsapi.config.Roles.ROLE_LEARNERS_RO
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.config.Roles.ROLE_LEARNERS_UI
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.integration.wiremock.LRSApiExtension.Companion.lrsApiMock
-import uk.gov.justice.digital.hmpps.learnerrecordsapi.models.db.MatchEntity
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.models.lrsapi.response.Learner
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.models.request.Gender
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.models.request.LearnersRequest
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.models.response.LRSResponseType
 import uk.gov.justice.digital.hmpps.learnerrecordsapi.models.response.LearnersResponse
-import uk.gov.justice.digital.hmpps.learnerrecordsapi.repository.MatchRepository
 import uk.gov.justice.hmpps.sqs.HmppsQueueService
 import uk.gov.justice.hmpps.sqs.MissingQueueException
 import uk.gov.justice.hmpps.sqs.audit.HmppsAuditEvent
@@ -32,9 +28,6 @@ class LearnersResourceIntTest : IntegrationTestBase() {
 
   @Autowired
   protected lateinit var objectMapper: ObjectMapper
-
-  @Autowired
-  lateinit var matchRepository: MatchRepository
 
   private val auditQueue by lazy {
     hmppsQueueService.findByQueueId("audit") ?: throw MissingQueueException("HmppsQueue audit not found")
@@ -381,57 +374,6 @@ class LearnersResourceIntTest : IntegrationTestBase() {
       )
 
       assertThat(receivedEvent.what).isEqualTo("SEARCH_LEARNER_BY_DEMOGRAPHICS")
-      assertThat(receivedEvent.who).isEqualTo("TestUser")
-      assertThat(receivedEvent.service).isEqualTo("hmpps-learner-records-api")
-      assertThat(receivedEvent.`when`).isBeforeOrEqualTo(Instant.now())
-    }
-
-    @Test
-    fun `should emit an event that request is received for find Learning Events by Nomis ID `() {
-      lrsApiMock.stubLearningEventsLinkedMatchFull()
-
-      auditSqsClient.purgeQueue(
-        PurgeQueueRequest.builder()
-          .queueUrl(auditQueueUrl)
-          .build(),
-      )
-
-      val requestJson = JSONObject()
-      requestJson.put("nomisId", "123456")
-
-      matchRepository.save(
-        MatchEntity(
-          null,
-          "123456",
-          "1234567890",
-          "Some Given Name",
-          "Some Family Name",
-          null,
-          Gender.MALE.toString(),
-        ),
-      )
-
-      webTestClient.post()
-        .uri("/learner-events/nomisId")
-        .headers(setAuthorisation(roles = listOf(ROLE_LEARNERS_RO)))
-        .header("X-Username", "TestUser")
-        .bodyValue(requestJson)
-        .accept(MediaType.parseMediaType("application/json"))
-        .exchange()
-        .expectStatus()
-        .is2xxSuccessful
-        .expectBody()
-        .returnResult()
-        .responseBody
-
-      val receivedEvent = objectMapper.readValue(
-        auditSqsClient.receiveMessage(
-          ReceiveMessageRequest.builder().queueUrl(auditQueueUrl).build(),
-        ).get().messages()[0].body(),
-        HmppsAuditEvent::class.java,
-      )
-
-      assertThat(receivedEvent.what).isEqualTo("SEARCH_LEARNER_EVENTS_BY_NOMISID")
       assertThat(receivedEvent.who).isEqualTo("TestUser")
       assertThat(receivedEvent.service).isEqualTo("hmpps-learner-records-api")
       assertThat(receivedEvent.`when`).isBeforeOrEqualTo(Instant.now())
